# nao_backend/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:5001",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read_pets": "read your pets",
        "write_pets": "modify pets in your account"
      }
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/organizations/": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "description": "Get all organizations",
        "parameters": [],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/Organizations"
            },
            "description": "Get all organizations"
          },
          "203": {
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "string",
                  "example": "No organization exists"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "No organization exists"
          }
        }
      }
    },
    "/organizations/getUser/{id}": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "description": "Get all user of an organization",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/UserOrganization"
            },
            "description": "All users of an organization"
          },
          "203": {
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "string",
                  "example": "Organization dosn't contain any user"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Organization dosn't contain any user"
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Organization dosn't exist"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Organization dosn't exist"
          }
        }
      }
    },
    "/organizations/userOrganisation/{id}": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "description": "Get all user of an organization",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/OrganizationUser"
            },
            "description": "All organization of an user"
          },
          "203": {
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "string",
                  "example": "User dosn't have access to any organization"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "User dosn't have access to any organization"
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "User dosn't exist"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "User dosn't exist"
          }
        }
      }
    },
    "/organizations/{name}": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "description": "Get organization by name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/Organizations"
            },
            "description": "Get organization by name"
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Orga dosn't exists"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Orga dosn't exists"
          }
        }
      }
    },
    "/organizations/addOrganization": {
      "post": {
        "tags": [
          "Organizations"
        ],
        "description": "Create organization",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "organizationName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/CreateOrga"
            },
            "description": "Created new Orga"
          },
          "409": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Orga already exists"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Orga already exists"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error creating Orga"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Error creating Orga"
          }
        }
      }
    },
    "/organizations/addUser": {
      "post": {
        "tags": [
          "Organizations"
        ],
        "description": "Add user to an organization",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "organizationId": {
                  "example": "any"
                },
                "organizationName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "string",
                  "example": "Created new User"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Created new User"
          },
          "409": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "User already exists"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "User already exists"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error deleted User"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Error deleted User"
          }
        }
      }
    },
    "/organizations/delete/user": {
      "delete": {
        "tags": [
          "Organizations"
        ],
        "description": "Add user to an organization",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "organizationId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "string",
                  "example": "Deleted User"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Deleted User"
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "User dosn't exists"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "User dosn't exists"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error deleted User"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Error deleted User"
          }
        }
      }
    },
    "/organizations/delete/{name}": {
      "delete": {
        "tags": [
          "Organizations"
        ],
        "description": "Delete organization",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "string",
                  "example": "Deleted organization"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Deleted organization"
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Organization dosn't exists"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Organization dosn't exists"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error deleted Orga"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Error deleted Orga"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Organizations": {
      "type": "object",
      "properties": {
        "result": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 36
              },
              "organizationName": {
                "type": "string",
                "example": "Epitech"
              }
            }
          }
        }
      }
    },
    "UserOrganization": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "example": 104
          },
          "organizationId": {
            "type": "number",
            "example": 38
          },
          "organizationName": {
            "type": "string",
            "example": "Adapei"
          }
        }
      }
    },
    "OrganizationUser": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "example": 104
          },
          "organizationId": {
            "type": "number",
            "example": 38
          },
          "organizationName": {
            "type": "string",
            "example": "Adapei"
          }
        }
      }
    },
...
    "CreateOrga": {
      "type": "object",
      "properties": {
        "organizationName": {
          "type": "string",
          "example": "Epi"
        }
      }
    },
...
  }
}